move pattern tests into their own file
